fbs_fastapi/requirements.txt
# FBS FastAPI Requirements
# Modernized dependencies for FBS FastAPI migration

# Core FastAPI
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0  # PostgreSQL async driver
alembic>=1.12.0,<2.0.0   # Database migrations

# Pydantic & Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
PyJWT>=2.8.0,<3.0.0  # JSON Web Token implementation
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# HTTP Client
httpx>=0.25.0,<1.0.0
aiohttp>=3.9.0,<4.0.0

# Caching
redis>=4.6.0,<6.0.0
aioredis>=2.0.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
structlog>=23.2.0,<24.0.0

# File Handling
aiofiles>=23.2.0,<24.0.0
python-magic>=0.4.27,<1.0.0

# Data Processing
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Testing (Development)
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
httpx>=0.25.0,<1.0.0

# Code Quality (Development)
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.7.0,<2.0.0

# Documentation (Development)
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.4.0,<10.0.0

# Monitoring (Optional)
sentry-sdk[fastapi]>=1.38.0,<2.0.0

# Legacy Compatibility (for gradual migration)
xmlrpc-client>=1.0.0  # For Odoo integration

# Development Server
watchfiles>=0.21.0,<1.0.0
